backend/requirements.txt
setuptools>=70.0.0
wheel>=0.42.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
anthropic>=0.32.0
duckdb>=1.0.0
pandas>=2.1.4
numpy>=1.26.0
boto3>=1.34.0
reportlab>=4.0.0
jinja2>=3.1.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
python-dotenv>=1.0.0
huggingface_hub>=0.20.0
requests>=2.31.0
mangum>=0.17.0
pytest>=7.4.0
mangum>=0.17.0
pytest>=7.4.0
